refactor inputs into seprate components
